requirements.txt
pytest
langchain
langchain-google-genai
pymongo>=4.7
python-dotenv
streamlit
numpy